pwawip

+++ components/userlist.vue
@@ -4,48 +4,23 @@
       <Toolbar class="mb-4">
         <template #start>
           <Button label="New" icon="pi pi-plus" class="p-button-success mr-2" @click="openNew" />
-          <Button
-            label="Delete"
-            icon="pi pi-trash"
-            class="p-button-danger"
-            @click="confirmDeleteSelected"
-            :disabled="!selectedUser || !selectedUser.length"
-          />
+          <Button label="Delete" icon="pi pi-trash" class="p-button-danger" @click="confirmDeleteSelected"
+            :disabled="!selectedUser || !selectedUser.length" />
         </template>
 
         <template #end>
-          <input
-            class="pa-4"
-            type="file"
-            ref="file"
-            style="display: none"
-            v-on:change="FileUpload($event)"
-          />
+          <input class="pa-4" type="file" ref="file" style="display: none" v-on:change="FileUpload($event)" />
 
           <Button v-bind="$attrs" v-on="$listeners" label="Import" @click="$refs.file.click()"></Button>
-          <Button
-            label="Export"
-            icon="pi pi-upload"
-            class="p-button-help mx-2"
-            @click="exportCSV($event)"
-          />
+          <Button label="Export" icon="pi pi-upload" class="p-button-help mx-2" @click="exportCSV($event)" />
         </template>
       </Toolbar>
 
-      <DataTable
-        ref="dt"
-        :value="userList"
-        :selection.sync="selectedUser"
-        data-key="id"
-        :paginator="true"
-        :rows="10"
+      <DataTable ref="dt" :value="userList" :selection.sync="selectedUser" data-key="id" :paginator="true" :rows="10"
         :filters="filters"
         paginatorTemplate="FirstPageLink PrevPageLink PageLinks NextPageLink LastPageLink CurrentPageReport RowsPerPageDropdown"
-        :rowsPerPageOptions="[5,10,25]"
-        currentPageReportTemplate="Showing {first} to {last} of {totalRecords} Users"
-        responsiveLayout="scroll"
-        :key="updateUserKey"
-      >
+        :rowsPerPageOptions="[5, 10, 25]" currentPageReportTemplate="Showing {first} to {last} of {totalRecords} Users"
+        responsiveLayout="scroll" :key="updateUserKey">
         <template #header>
           <div class="table-header flex flex-column md:flex-row md:justify-content-between">
             <h5 class="mb-2 md:m-0 md:align-self-center">Manage Users</h5>
@@ -56,57 +31,31 @@
           </div>
         </template>
 
-        <Column selectionMode="multiple" :styless="{width: '3rem'}" :exportable="false"></Column>
-        <Column field="id" header="Employee Id" :sortable="true" :styles="{'min-width':'12rem'}"></Column>
-        <Column field="fullName" header="Name" :sortable="true" :styles="{'min-width':'16rem'}"></Column>
-        <Column field="mobile" header="Mobile" :sortable="true" :styles="{'min-width':'8rem'}"></Column>
-        <Column field="emailid" header="Email" :sortable="true" :styles="{'min-width':'10rem'}"></Column>
-        <Column field="status" header="Status" :sortable="true" :styles="{'min-width':'10rem'}"></Column>
-        <Column
-          field="createdAt"
-          header="createdAt"
-          :sortable="true"
-          :styles="{'min-width':'10rem'}"
-        ></Column>
-        <Column :exportable="false" :styles="{'min-width':'8rem'}">
+        <Column selectionMode="multiple" :styless="{ width: '3rem' }" :exportable="false"></Column>
+        <Column field="id" header="Employee Id" :sortable="true" :styles="{ 'min-width': '12rem' }"></Column>
+        <Column field="fullName" header="Name" :sortable="true" :styles="{ 'min-width': '16rem' }"></Column>
+        <Column field="mobile" header="Mobile" :sortable="true" :styles="{ 'min-width': '8rem' }"></Column>
+        <Column field="emailid" header="Email" :sortable="true" :styles="{ 'min-width': '10rem' }"></Column>
+        <Column field="status" header="Status" :sortable="true" :styles="{ 'min-width': '10rem' }"></Column>
+        <Column field="createdAt" header="createdAt" :sortable="true" :styles="{ 'min-width': '10rem' }"></Column>
+        <Column :exportable="false" :styles="{ 'min-width': '8rem' }">
           <template #body="slotProps">
-            <Button
-              icon="pi pi-pencil"
-              class="p-button-rounded p-button-success mr-2"
-              @click="editUser(slotProps.data)"
-            />
-            <Button
-              icon="pi pi-trash"
-              class="p-button-rounded p-button-warning"
-              @click="confirmDeleteUser(slotProps.data)"
-            />
+            <Button icon="pi pi-pencil" class="p-button-rounded p-button-success mr-2"
+              @click="editUser(slotProps.data)" />
+            <Button icon="pi pi-trash" class="p-button-rounded p-button-warning"
+              @click="confirmDeleteUser(slotProps.data)" />
           </template>
         </Column>
       </DataTable>
     </div>
 
-    <Dialog
-      :visible.sync="userDialog"
-      :style="{width: '450px'}"
-      header="User Details"
-      :modal="true"
-      class="p-fluid"
-    >
+    <Dialog :visible.sync="userDialog" :style="{ width: '450px' }" header="User Details" :modal="true" class="p-fluid">
       <userForm @hideDialog="hideDialog" @saveUser="saveUser" :userAction="action" :user="user" />
     </Dialog>
 
-    <Dialog
-      :visible.sync="deleteUserDialog"
-      :styles="{width: '450px'}"
-      header="Confirm"
-      :modal="true"
-    >
-      <confirmationDialog
-        :message="message"
-        :deleteAll="deleteAll"
-        @confirmAction="deleteUser"
-        @closeAction="deleteUserDialog=false"
-      />
+    <Dialog :visible.sync="deleteUserDialog" :styles="{ width: '450px' }" header="Confirm" :modal="true">
+      <confirmationDialog :message="message" :deleteAll="deleteAll" @confirmAction="deleteUser"
+        @closeAction="deleteUserDialog = false" />
     </Dialog>
   </div>
 </template>
@@ -161,9 +110,9 @@ export default {
       console.log("file.........", file, results);
       let importData = results?.data || [];
       importData.pop();
-      const userCreatedData=await this.bulkCreateUser([...importData]);
-      if(userCreatedData){
-      await  this.fetchUser();
+      const userCreatedData = await this.bulkCreateUser([...importData]);
+      if (userCreatedData) {
+        await this.fetchUser();
       }
     },
     openNew() {
@@ -224,8 +173,9 @@ export default {
     },
     async deleteUser() {
       let users = [];
-      users = this.selectedUser || users.push(this.user);
-      await this.deleteUserFunc(users);
+      if (this.user) users.push(this.user)
+      const userObject = this.selectedUser || users;
+      await this.deleteUserFunc(userObject);
       this.user = {};
       this.selectedUser = [];
     },
@@ -238,6 +188,7 @@ export default {
       this.deleteUserDialog = true;
     },
     async deleteUserFunc(users) {
+      console.log("userrrrrrrrrrrrr", users)
       const deletedUser = await this.deleteUsers(users);
       if (deletedUser) {
         this.userList = this.userList.filter(obj => {
